Embed creator

+++ commands/Util/Embed.js
@@ -16,7 +16,7 @@ module.exports.run = async (bot,message,args,cmd) => {
 
     if(isNaN(num)) return new _NoticeEmbed(Colors.ERROR, "Invalid channel - Please specify a valid channel (either #channel-name or id)").send(message.channel);
 
-    num = parseFloat(num);
+    num = parseInt(num);
 
     let channel = message.guild.channels.cache.get(num);
